Major Changes
Made a bunch of css changes and I am opting out of the snap to window functionality. Its not in the original site so why build it. Still have 28 hero's and a footer to go. I figure that can be split up to 5 heros a day over the next 5 days.