added extra padding between sections and carousels.
Really I was testing out my new keyboard and seeing how typing on the Keychron 4 would feel in my code flow. I gotta say I'm not dissapointed. Perhaps there is a more tactile experience with the other keyboards but as of right now I'm pretty happy with this one. Lets see the click is pretty satisfying.